rename jobs to fw model steps
